+++ xknx/devices/expose_sensor.py
@@ -88,9 +88,9 @@ class ExposeSensor(Device):
 
     def __str__(self) -> str:
         """Return object as readable string."""
-        return '<ExposeSensor name="{}" ' 'sensor="{}" value="{}" unit="{}"/>'.format(
+        return '<ExposeSensor name="{}" sensor={} value={} unit="{}"/>'.format(
             self.name,
             self.sensor_value.group_addr_str(),
-            self.resolve_state(),
+            self.resolve_state().__repr__(),
             self.unit_of_measurement(),
         )